Changed doubles to floats as those are easier to work with for the GPU

+++ LSH_cpp2/lsh.cpp
@@ -10,19 +10,19 @@
 #include <sys/time.h>
 #endif
 
-inline double rand_minus1_to_1() {
-    return ((double)rand() / (RAND_MAX/2)) - 1;
+inline float rand_minus1_to_1() {
+    return ((float)rand() / (RAND_MAX/2)) - 1;
 }
 
-void fill_point_arrays(int nPoints1, int nPoints2, int vectorLength, double noiseScale,
-                       double* points1, double* points2);
+void fill_point_arrays(int nPoints1, int nPoints2, int vectorLength, float noiseScale,
+                       float* points1, float* points2);
 
-void fill_hyperplanes(int nPlanes, int vectorLength, double* hyperplanes);
+void fill_hyperplanes(int nPlanes, int vectorLength, float* hyperplanes);
 
 void calculate_hash_values(// input
-                           int vectorLength, int nPoints1, int nPlanes, double* points,
+                           int vectorLength, int nPoints1, int nPlanes, float* points,
                            // output
-                           double* hyperplanes, int* indexGroupMap);
+                           float* hyperplanes, int* indexGroupMap);
 
 void organize_points_into_groups(// input
                                  int nPoints1, int nBoxes,
@@ -34,16 +34,16 @@ void organize_points_into_groups(// input
 
 void calculate_indexGroupMap(// input
                              int vectorLength, int numTables, int nPlanes,
-                             int nPoints, double* points,
-                             double* hyperplanes, int hyperplanesTableLen,
+                             int nPoints, float* points,
+                             float* hyperplanes, int hyperplanesTableLen,
                              int indexGroupMapTableLen,
                              // output
                              int* indexGroupMap);
 
 void construct_lsh_tables(// input
                           int vectorLength, int numTables, int nPlanes, int nBoxes,
-                          int nPoints1, double* points1,
-                          double* hyperplanes, int hyperplanesTableLen,
+                          int nPoints1, float* points1,
+                          float* hyperplanes, int hyperplanesTableLen,
                           // preallocated temporary storage
                           int indexGroupMapTableLen, int* indexGroupMap,
                           // output
@@ -62,10 +62,10 @@ int find_potential_matches(// inputs
 
 void match_points(// inputs
                   int vectorLength, int nPoints2,
-                  double* points1, double* points2,
+                  float* points1, float* points2,
                   int* potentialMatches, int* potentialMatchesIndices, int* potentialMatchesLengths,
                   // outputs
-                  int* lshMatches, double* bestMatchDists);
+                  int* lshMatches, float* bestMatchDists);
 
 
 int double_int_arr_size(int** arr, int curSize) {
@@ -85,14 +85,14 @@ int main() {
     const int nPoints1 = 10000;     // number of points in the first dataset
     const int nPoints2 = nPoints1;  // number of points in the second dataset
     const int vectorLength = 128;
-    const double noiseScale = 0.3;
+    const float noiseScale = 0.3;
     const int numTables = 8;
 
     // -- Generate some random points and similar random points to match with --
 
     // points to be matched
-    double* points1 = (double*)malloc(numTables * nPoints1 * vectorLength * sizeof(double));
-    double* points2 = (double*)malloc(numTables * nPoints2 * vectorLength * sizeof(double));
+    float* points1 = (float*)malloc(numTables * nPoints1 * vectorLength * sizeof(float));
+    float* points2 = (float*)malloc(numTables * nPoints2 * vectorLength * sizeof(float));
 
     const int nPlanes = (int)log2(nPoints1);
 
@@ -103,11 +103,11 @@ int main() {
     // allocate hyperplanes
     const int hyperplanesTableLen = nPlanes * vectorLength;
     const int hyperplanesLen = numTables * hyperplanesTableLen;
-    double* hyperplanes = (double*)malloc(hyperplanesLen * sizeof(double));
+    float* hyperplanes = (float*)malloc(hyperplanesLen * sizeof(float));
 
     // fill all hyperplanes
     for (int table = 0; table < numTables; table++) {
-        double* hyperplanes2 = hyperplanes + table * hyperplanesTableLen;
+        float* hyperplanes2 = hyperplanes + table * hyperplanesTableLen;
         // create normalized hyperplanes represented by vectors of euclidean size 1
         fill_hyperplanes(nPlanes, vectorLength, hyperplanes2);
     }
@@ -200,7 +200,7 @@ int main() {
     int* lshMatches = (int*)malloc(nPoints2 * sizeof(int));
     memset(lshMatches, -1, nPoints2 * sizeof(int));
     // holds the best match for each point we're finding a match for
-    double* bestMatchDists = (double*)malloc(nPoints2 * sizeof(double));
+    float* bestMatchDists = (float*)malloc(nPoints2 * sizeof(float));
     for (int i = 0; i < nPoints2; i++) {
         bestMatchDists[i] = 1e10;
     }
@@ -240,18 +240,18 @@ int main() {
     free(potentialMatchesLengths);
 }
 
-void fill_point_arrays(int nPoints1, int nPoints2, int vectorLength, double noiseScale,
-                       double* points1, double* points2)
+void fill_point_arrays(int nPoints1, int nPoints2, int vectorLength, float noiseScale,
+                       float* points1, float* points2)
 {
     for (int i = 0; i < ((nPoints1 < nPoints2) ? nPoints1 : nPoints2); i++) {
-        double size1 = 0;
-        double size2 = 0;
+        float size1 = 0;
+        float size2 = 0;
         for (int j = 0; j < vectorLength; j++) {
             // points1[i][:] = random array with values between -1 and 1
-            double rn = rand_minus1_to_1();
+            float rn = rand_minus1_to_1();
             points1[i*vectorLength + j] = rn;
             // points2[i][:] = points1[i][:] + noise
-            double noise = noiseScale * rand_minus1_to_1();
+            float noise = noiseScale * rand_minus1_to_1();
             points2[i*vectorLength + j] = rn + noise;
             // Keep track of euclidean lengths to normalize after loop
             size1 += points1[i*vectorLength + j] * points1[i*vectorLength + j];
@@ -269,10 +269,10 @@ void fill_point_arrays(int nPoints1, int nPoints2, int vectorLength, double nois
     }
 }
 
-void fill_hyperplanes(int nPlanes, int vectorLength, double* hyperplanes) {
+void fill_hyperplanes(int nPlanes, int vectorLength, float* hyperplanes) {
 
     for (int i = 0; i < nPlanes; i++) {
-        double size = 0;
+        float size = 0;
         for (int j = 0; j < vectorLength; j++) {
             // hyperplanes[i][:] = random array with values between -1 and 1
             hyperplanes[i*vectorLength + j] = rand_minus1_to_1();
@@ -282,7 +282,7 @@ void fill_hyperplanes(int nPlanes, int vectorLength, double* hyperplanes) {
 
         
         size = sqrt(size);
-        double invsize = 1/size;
+        float invsize = 1/size;
         // normalize the vectors
         for (int j = 0; j < vectorLength; j++) {
             hyperplanes[i * vectorLength + j] *= invsize;
@@ -291,19 +291,19 @@ void fill_hyperplanes(int nPlanes, int vectorLength, double* hyperplanes) {
 }
 
 void calculate_hash_values(// input
-                           int vectorLength, int nPoints1, int nPlanes, double* points,
+                           int vectorLength, int nPoints1, int nPlanes, float* points,
                            // output
-                           double* hyperplanes, int* indexGroupMap)
+                           float* hyperplanes, int* indexGroupMap)
 {
     // - Calculate hash values, keep track of sizes of each group -
     for (int i = 0; i < nPoints1; i++) {
-        double* point = &points[i * vectorLength];
+        float* point = &points[i * vectorLength];
         int hashcode = 0;  //  hashcode will be the group index
         // calculate hash value of the i'th point, store resut in indexGroupMap
-        double* hplane = hyperplanes;  // first hyperplane
+        float* hplane = hyperplanes;  // first hyperplane
         for (int j = 0; j < nPlanes; j++) {
             // calculate point * hplane
-            double vecMul = 0;
+            float vecMul = 0;
             for (int k = 0; k < vectorLength; k++) {
                 vecMul += point[k] * hplane[k];
             }
@@ -356,14 +356,14 @@ void organize_points_into_groups(// input
 
 void calculate_indexGroupMap(// input
                              int vectorLength, int numTables, int nPlanes,
-                             int nPoints, double* points,
-                             double* hyperplanes, int hyperplanesTableLen,
+                             int nPoints, float* points,
+                             float* hyperplanes, int hyperplanesTableLen,
                              int indexGroupMapTableLen,
                              // output
                              int* indexGroupMap)
 {
     for (int table = 0; table < numTables; table++) {
-        double* hyperplanes2 = hyperplanes + table * hyperplanesTableLen;
+        float* hyperplanes2 = hyperplanes + table * hyperplanesTableLen;
         int* indexGroupMap2 = indexGroupMap + table * indexGroupMapTableLen;
         // - Match points -
         calculate_hash_values(vectorLength, nPoints, nPlanes,
@@ -374,8 +374,8 @@ void calculate_indexGroupMap(// input
 
 void construct_lsh_tables(// input
                           int vectorLength, int numTables, int nPlanes, int nBoxes,
-                          int nPoints1, double* points1,
-                          double* hyperplanes, int hyperplanesTableLen,
+                          int nPoints1, float* points1,
+                          float* hyperplanes, int hyperplanesTableLen,
                           int indexGroupMapTableLen, int* indexGroupMap,
                           // output
                           int* groupArray, int groupArrayTableLen,
@@ -466,15 +466,15 @@ int find_potential_matches(// inputs
 
 void match_points(// inputs
                   int vectorLength, int nPoints2,
-                  double* points1, double* points2,
+                  float* points1, float* points2,
                   int* potentialMatches, int* potentialMatchesIndices, int* potentialMatchesLengths,
                   // outputs
-                  int* lshMatches, double* bestMatchDists)
+                  int* lshMatches, float* bestMatchDists)
 {
     // printf("new way:");
     for (int i = 0; i < nPoints2; i++) {
         bool changed = false;
-        double bestFitDist = bestMatchDists[i];
+        float bestFitDist = bestMatchDists[i];
         int match = -1;
         int jStart = potentialMatchesIndices[i];
         int jEnd = potentialMatchesIndices[i] + potentialMatchesLengths[i];
@@ -485,9 +485,9 @@ void match_points(// inputs
             int idx = potentialMatches[j];
             // printf(" - %d\n", idx);
             // diff = sum((points2[i][:] - points1[idx][:]) .^ 2);
-            double diff = 0;
+            float diff = 0;
             for (int k = 0; k < vectorLength; k++) {
-                double tmp = points2[i * vectorLength + k] - points1[idx * vectorLength + k];
+                float tmp = points2[i * vectorLength + k] - points1[idx * vectorLength + k];
                 diff += tmp * tmp;
             }
             // check if distance is the lowest distance so far
